fix(lock): indicate invalid desciptor with -1
Previously, 0 was used to indicate an unset descriptor, even though this
would be a valid one (stdin most of the time).
Now -1 is used consistently to indicate that the descriptor is not set.

Code using the `_fd` field was adjusted to use `_errno` instead, as it
is the field that truly indicates an error condition.