refactor(lock): Lock as an object
This commit is in preparation for the removing side-effects currently
present, namely the creation of a File.

As a first step, macros were removed to reduce code complexity. This
allowed for a more explicit interface which makes clear that a file
is created in the process (if not present), which is subsequently
locked.

The API should eventually allow to just lock and unlock a file, in a
platform independent manner.

Blocking and non-blocking modes are now controlled through an enum,
which should make configurable lock-types easier.