refactor(lock): don't enforce file-creation
To my mind the essential feature of the implementation is the ability
to lock any existing file, as identified by file descriptor. It should
be left to the user to create and optionally delete the file, which
allows for more more flexibility. Cargo for instance could use its own
fingerprint files for its collaborative locking scheme.

Tests were adjusted to create the required file and take care of
cleanup as well.

The API was changed to fully embrace an object oriented paradigm, also
because it seems to be a better fit to the standard library.