docs(lock): updated to represent current state
This involves an attempt to remove some duplication within the examples,
and be clear about the kind of lock we actually are, i.e. a write lock.

The latter hints at a desireable feature, namely the ability to obtain
a read-lock as well.

Before that can happen though, one should probably try to get real
tests going to see it actually works among multiple processes.