tests(lock): multi-process lock test
A dedicated test was added which calls itself to make the attempt and
lock the same file from multiple processes. The test called first
controls its second invocation via environment variables.

However, it shows that the second process can obtain a WRITE lock even
though the first one already acquired one.

Something doesn't work here, and it seems to be similar on linux as
well (even though this will be tested only when pushed so travis
can have a go).