feat(filelock): initial implementation
The convenience wrapper doesn't have tests yet, and won't behave
correctly if a lock could not be obtained and it is dropped in
case we didn't create the file.

We would need a way to know if we created the file or not. Otherwise,
there is no safe, non-racy way of file removal ... .